requirements.txt
streamlit==1.40.2  
mistralai==1.1.0    
tenacity==9.0.0     
python-dotenv==1.0.1